fix(react-intlayer, next-intlayer): add tree shaking for both react-intlayer, next-intlayer & move format hooks to format folder
