+++ packages/react-intlayer/package.json
@@ -41,10 +41,20 @@
       "require": "./dist/cjs/index.cjs",
       "import": "./dist/esm/index.mjs"
     },
+    "./format": {
+      "types": "./dist/types/client/format/index.d.ts",
+      "require": "./dist/cjs/client/format/index.cjs",
+      "import": "./dist/esm/client/format/index.mjs"
+    },
     "./server": {
       "types": "./dist/types/server/index.d.ts",
       "require": "./dist/cjs/server/index.cjs",
       "import": "./dist/esm/server/index.mjs"
+    },
+    "./server/format": {
+      "types": "./dist/types/server/format/index.d.ts",
+      "require": "./dist/cjs/server/format/index.cjs",
+      "import": "./dist/esm/server/format/index.mjs"
     },
     "./editor": {
       "types": "./dist/types/editor/index.d.ts",
