+++ packages/react-intlayer/src/client/format/useFormatter.ts
@@ -1,8 +1,9 @@
+'use client';
+
 import { useCompact } from './useCompact';
 import { useCurrency } from './useCurrency';
 import { useDate } from './useDate';
 import { useList } from './useList';
-import { useListFormatter } from './useListFormatter';
 import { useNumber } from './useNumber';
 import { usePercentage } from './usePercentage';
 import { useRelativeTime } from './useRelativeTime';
@@ -17,7 +18,6 @@ export const useFormatter = () => {
   const date = useDate();
   const relativeTime = useRelativeTime();
   const list = useList();
-  const listFormatter = useListFormatter();
 
   return {
     currency,
@@ -28,6 +28,5 @@ export const useFormatter = () => {
     date,
     relativeTime,
     list,
-    listFormatter,
   };
 };
